Initial ion commit. This will be fleshed out into a fully functional shell for redox.
Changes to be committed:
	modified:   filesystem/apps/terminal/terminal.rs
	copied:     filesystem/apps/terminal/terminal.rs -> libredox/src/ion/command.rs
	new file:   libredox/src/ion/mod.rs
	modified:   libredox/src/lib.rs
